+++ drivacy/primitives/secret_sharing_test.cc
@@ -31,7 +31,7 @@ uint64_t Test(uint64_t value, uint64_t numparties) {
 int main() {
   for (int i = 0; i < 100; i++) {
     uint64_t value = std::rand() % drivacy::primitives::PRIME;
-    uint64_t numparties = std::rand() % 5 + 2;  // [2, 7)
+    uint32_t numparties = std::rand() % 5 + 2;  // [2, 7)
     uint64_t reconstructed = Test(value, numparties);
     if (value != reconstructed) {
       std::cout << "Test failed!" << std::endl;
